Rename to database.js
