+++ api/controllers/artists.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var context = require('../../index.js').getContext();
+var context = require('../../database.js').getContext();
 var artists = context.component('dsap').module('artists');
 
 module.exports = {
